# [2.0] Improvements in Uri::cleanPath()