@frankmayer 2017-07-26
LGTM.
And nice catch of the bugs. The only question with those getting fixed, is that they might break stuff, if people are expecting the wrong behavior.